Create READ ME
webapptest
Test for interviews
This test is designed to give an overview of your capabilities as a java developer. The ultimate goal of the test is spelled out in the main method of src/com/proquest/interview/phonebook/PhoneBookImpl.
Implement the TODO comments to the best of your ability. In addition, you should be looking at how to improve this program. The DatabaseUtil class should not need to be significantly refactored, although you may do so as long as you do not destroy it's functionality.
You can download this project, make your changes, zip/tar it up and send it back to us, or if you want to impress us, fork it, put your changes in, and send along the url.
This test should take less than an hour.
Thank you.

*************************** Comment of my implementation **************************************

I created a properties file to be able to dynamically configure the connection, but at the moment and commented on the functionality not to overturn the mother code.
I also created a custome class to capture the exception. "CutomIllegalArgumentException"

in the testing phase there are several more methods to create a dynamic query, created constant class for all the words that can be recalled by several parts.

there is a class "DBConnectionUtil" : creating the connection but not being used is only commented on to make the idea of how I would operate.

once the project has been run in run, the initialization of the DB must be commented, as the records are already inserted and so avoid the exception that can arise